+++ app/javascript/components/game/map/MapHexOverlay.tsx
@@ -9,10 +9,11 @@ interface MapHexOverlayProps {
   hex: Hex;
   selectCallback: (x: number, y: number) => void,
   shaded: HexOpenType;
+  firingSmoke: boolean;
 }
 
 export default function MapHexOverlay({
-  hex, selectCallback, shaded
+  hex, selectCallback, shaded, firingSmoke,
 }: MapHexOverlayProps) {
   const shadedStyle = { fill: shadedHexColor }
   const falseClosedStyle = { fill: falseShadedHexColor }
@@ -38,7 +39,7 @@ export default function MapHexOverlay({
       let decoration = shaded === hexOpenType.All ? "A" : shaded
       if (decoration === 0.5) { decoration = "½" }
       circle = (
-        <g transform={ hex.map.rotated ? `rotate(90 ${x + xoffset} ${y - yoffset})` : "" }>
+        <g transform={ hex.map.rotated ? `rotate(90 ${x + xoffset} ${y - yoffset})` : "" } >
           <path d={circlePath(new Coordinate(x + xoffset, y - yoffset), 30)}
                 style={{ fill: "rgba(0,0,0,0.3)" }} />
           <text x={x + xoffset} y={y - yoffset + 15} fontSize={56} textAnchor="middle"
@@ -49,7 +50,7 @@ export default function MapHexOverlay({
       )
     }
     return (
-      <g>
+      <g className={firingSmoke ? "smoke-cursor" : ""} >
         { circle }
         <polygon points={hex.hexCoords} style={style}
                  onClick={() => open ? selectCallback(hex.coord.x, hex.coord.y) : {}}